moved edit image file around
